+++ tests/testthat/test-provider-google.R
@@ -59,6 +59,27 @@ test_that("can search web pages", {
   test_tool_web_search(chat_fun, google_tool_web_search())
 })
 
+test_that("can combine built-in and user tools", {
+  provider <- chat_google_gemini_test()$get_provider()
+
+  regular_tool <- tool(
+    function(x) x,
+    "Return `x`",
+    arguments = list(x = type_number("x"))
+  )
+
+  body <- chat_body(
+    provider,
+    stream = TRUE,
+    turns = list(Turn("user", "hi")),
+    tools = list(regular_tool, google_tool_web_search())
+  )
+
+  expect_length(body$tools, 2)
+  expect_named(body$tools[[1]], "functionDeclarations")
+  expect_named(body$tools[[2]], "google_search")
+})
+
 test_that("can use images", {
   vcr::local_cassette("google-image")
   chat_fun <- chat_google_gemini_test
@@ -237,7 +258,7 @@ test_that("batch_status waits for responsesFile after SUCCEEDED", {
   no_file <- batch_status(provider, returned_batch)
   expect_true(no_file$working)
 
-  returned_batch$response = list(responsesFile = "files/abc123")
+  returned_batch$response <- list(responsesFile = "files/abc123")
 
   with_file <- batch_status(
     provider,